Added an option to the setup file to specify the number of worker threads to create.

+++ src/Threading/JobDispatcher.cpp
@@ -12,15 +12,15 @@ namespace AV{
     std::queue<Worker*> JobDispatcher::workersQueue;
     std::deque<JobDispatcher::JobEntry> JobDispatcher::jobQueue;
     uint64_t JobDispatcher::jobCount = 0;
-    
+
     std::mutex JobDispatcher::waitMutex;
     std::unique_lock<std::mutex> JobDispatcher::waitLock;
     //Creating 4 manually because I'm not sure if this is the solution.
     std::vector<std::condition_variable> JobDispatcher::waitCv(4);
-    
+
     const JobId JobId::INVALID;
 
-    bool JobDispatcher::initialise(int numWorkers){
+    bool JobDispatcher::initialise(uint8 numWorkers){
         AV_INFO("Job Dispatcher creating {} threads", numWorkers);
         std::thread *t = 0;
         Worker* w = 0;
@@ -30,7 +30,7 @@ namespace AV{
             t = new std::thread(&Worker::run, w);
             threads.push_back(t);
         }
-        
+
         waitLock = std::unique_lock<std::mutex>(waitMutex);
 
         return true;
@@ -48,23 +48,23 @@ namespace AV{
 
             AV_INFO("Joined worker thread {}", i);
         }
-        
+
         threads.clear();
         workers.clear();
 
         return true;
     }
-    
+
     void JobDispatcher::endJob(JobId job){
         if(job == JobId::INVALID) return;
         //Check if a worker contains this job.
         //Otherwise check the job queue.
         //If nothing could be found then return straight away because there is no job with that id.
-        
+
         //Both need to be locked, as otherwise a race condition is possible.
         std::unique_lock<std::mutex> workersLock(workersMutex);
         std::unique_lock<std::mutex> jobLock(jobMutex);
-        
+
         Worker* targetW = 0;
         for(Worker* w : workers){
             if(w->runningJob(job)){
@@ -72,25 +72,25 @@ namespace AV{
                 break;
             }
         }
-        
+
         if(targetW != 0){
             //The job is being run by a worker. So I should wait for that worker.
             AV_INFO("Job {} being run by worker.", job.id());
-            
+
             //The mutexes can be unlocked, so they're only for the searching part.
             //If they're not unlocked the system will reach a deadlock in the addWorkerToQueue function.
             workersLock.unlock();
             jobLock.unlock();
-            
+
             waitCv[targetW->getWorkerId()].wait(waitLock);
             AV_INFO("Finished waiting.");
-            
+
             return;
         }else{
             //The job is not being run by a worker.
             //Search the queue now.
             AV_INFO("Job {} not being run by worker.", job.id());
-            
+
             auto jit = jobQueue.begin();
             while(jit != jobQueue.end()){
                 if((*jit).first == job) break;
@@ -110,7 +110,7 @@ namespace AV{
     JobId JobDispatcher::dispatchJob(Job *job){
         //Lock things up.
         std::unique_lock<std::mutex> workersLock(workersMutex);
-        
+
         //Increment the job count. The value it has now will be the id of this job.
         jobCount++;
         JobId jobId(jobCount);
@@ -124,7 +124,7 @@ namespace AV{
             cv = worker->getConditionVariable();
             cv->notify_one();
             AV_INFO("Notifying");
-            
+
             AV_INFO("Job {} going straight to worker.", jobCount);
 
             workersQueue.pop();
@@ -135,7 +135,7 @@ namespace AV{
             std::unique_lock<std::mutex> jobLock(jobMutex);
             jobQueue.push_back(jobEntry);
         }
-        
+
         return jobId;
     }
